Extended dagster repositories to include base, dev, and prod. Helmfile changes to load the correct repositories file when rendering the chart. Removing the new pipeline and moving the weekly_sample_refresh to the prod repositories code.
