+++ orchestration/dap_orchestration/pipelines.py
@@ -78,43 +78,3 @@ def refresh_data_all() -> None:
     ]
     write_outfiles(collected_outputs)
 
-
-@pipeline(
-    mode_defs=[local_mode, dev_mode, prod_mode, test_mode]
-)
-def refresh_sample_data() -> None:
-    collected_outputs = [
-        sample_transform_records(sample_extract_records())
-    ]
-    write_outfiles(collected_outputs)
-
-
-@weekly_schedule(
-    pipeline_name="refresh_sample_data",
-    start_date=datetime(2021, 9, 9),
-    execution_time=time(15, 00),
-    execution_timezone="US/Eastern",
-    mode="test_mode",
-    execution_day_of_week=1
-)
-def weekly_sample_refresh(date: datetime) -> dict[str, object]:
-    return {
-        "resources": {
-            "refresh_directory": {
-                "config": {"/refresh_output"}
-            }
-        },
-        "solids": {
-            "write_outfiles": {
-                "config": {
-                    "working_dir": ".."
-                }
-            },
-            "sample_extract_records": {
-                "config": {
-                    "pull_data_dictionaries": {"false"},
-                    "end_time": date.strftime("%Y-%m-%d %H")
-                }
-            }
-        }
-    }
