+++ orchestration/dap_orchestration/repositories/dev_repositories.py
@@ -1,0 +1,7 @@
+from dagster import PipelineDefinition, repository
+
+from dap_orchestration.repositories.base_repositories import all_jobs
+
+@repository
+def repositories() -> list[PipelineDefinition]:
+    return all_jobs()
