+++ orchestration/workspace.yaml
@@ -1,2 +1,2 @@
 load_from:
-  - python_file: repositories.py
+  - python_file: prod_repositories.py
